requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
djangorestframework==3.10.1
gunicorn==19.9.0
idna==2.8
omdb==0.10.1
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3